refactor: [#100] move high-level dvc library methods
These methods should not be in the generic DVC mod.
They filter files that belong to a library (media files).
The generic DVC mod should know nothing about librarian librariries and
namecodes specification.
