+++ src/nautilus_librarian/typer/commands/workflows/actions/delete_base_images_action.py
@@ -1,8 +1,10 @@
 from os import path, remove
 
+from nautilus_librarian.domain.dvc_diff_parser import (
+    extract_deleted_files_from_dvc_diff,
+)
 from nautilus_librarian.domain.dvc_services_api import DvcServicesApi
 from nautilus_librarian.domain.file_locator import file_locator
-from nautilus_librarian.mods.dvc.domain.utils import extract_deleted_files_from_dvc_diff
 from nautilus_librarian.mods.namecodes.domain.filename import Filename
 from nautilus_librarian.typer.commands.workflows.actions.action_result import (
     ActionResult,
