+++ src/nautilus_librarian/typer/commands/workflows/actions/dvc_pull_action.py
@@ -1,7 +1,7 @@
-from nautilus_librarian.domain.dvc_services_api import DvcServicesApi
-from nautilus_librarian.mods.dvc.domain.utils import (
+from nautilus_librarian.domain.dvc_diff_parser import (
     extract_added_and_modified_files_from_dvc_diff,
 )
+from nautilus_librarian.domain.dvc_services_api import DvcServicesApi
 from nautilus_librarian.typer.commands.workflows.actions.action_result import (
     ActionResult,
     ErrorMessage,
