+++ src/nautilus_librarian/typer/commands/workflows/actions/validate_filepaths_action.py
@@ -1,6 +1,8 @@
+from nautilus_librarian.domain.dvc_diff_parser import (
+    extract_list_of_new_or_renamed_files_from_dvc_diff_output,
+)
 from nautilus_librarian.domain.validate_filepaths import validate_filepath
 from nautilus_librarian.mods.dvc.domain.utils import (
-    extract_list_of_new_or_renamed_files_from_dvc_diff_output,
     get_new_filepath_if_is_a_renaming_dict,
 )
 from nautilus_librarian.typer.commands.workflows.actions.action_result import (
